feat(dialogs): Incorporate DialogDependencies
Dialog dependencies is a feature in

- [botbuilder-js](https://github.com/microsoft/botbuilder-js/blob/87d0ad1a889a38396e761f10614c78a4526590d1/libraries/botbuilder-dialogs/src/dialogSet.ts#L169),
- and in [botbuilder-dotnet](https://github.com/microsoft/botbuilder-dotnet/blob/bc17a5db2c717fec0d4c109d884bea88d29f75cb/libraries/Microsoft.Bot.Builder.Dialogs/DialogSet.cs#L147).

To maintain feature parity, and because this functionality is extremely
useful this PR brings it to `botbuilder-python`.

Tecnically in the `botbuilder-js` and dotnet implementations dialogs that
are found to conflict get stored with a different ID, but I am sure that
would likely be a breaking change (and IMO a very counter-intuitive way
of working).

If you are adding dialog dependencies and they conflict, the logic as it
was would be retained.

Note that, although by making it a `@runtime_checkable` Protocol with
DialogDependencies one can check `isinstance(object, DialogDependencies)`,
the recommendation from the python docs is:

"check against a runtime-checkable protocol can be surprisingly slow"
— [python typing](https://docs.python.org/3/library/typing.html#typing.Protocol)

For that reason we use `hasattr` and `callable` instead.

Also, note that protocol is avaialble from python 3.8 onwards,
which is already the minimum version for this library.

The purpose of using Iterable instead of List is just to keep it generic.
Since python has the concepts of `tuple`, `List`, etc and to keep it as a structural type.